--- backend/lores-node-axum/src/local_apps/coop_cloud_apps.rs
@@ -1,1 +1,0 @@
-pub use coop_cloud_docker_apps::{AppUrl, CoopCloudApp, find_coop_cloud_apps};
